requirements.txt
ja-ginza==5.1.3
langchain==0.1.9
spacy==3.6.1
